fix: move common component
