+++ frontend/src/core/components/Button/Button.tsx
@@ -1,7 +1,7 @@
 import { Button, SxProps, Theme, ButtonProps } from "@mui/material";
 import React from "react";
 
-import { rubik } from "@/app/fonts";
+import { rubik } from "@/core/utils/fonts";
 
 interface IButtonComponent {
   children?: React.ReactNode;
